Restructure part 5 - Move page components to containers folder - Change enrolment to auth - Fix import paths
