+++ src/app/app.module.ts
@@ -27,16 +27,16 @@ import { KafkaService } from './core/services/kafka.service'
 import { NotificationService } from './core/services/notification.service'
 import { SchedulingService } from './core/services/scheduling.service'
 import { StorageService } from './core/services/storage.service'
-import { ClinicalTasksPage } from './pages/clinical-tasks/clinical-tasks-page.component'
-import { EnrolmentPage } from './pages/enrolment/enrolment-page.component'
-import { AuthService } from './pages/enrolment/services/auth.service'
-import { FinishPage } from './pages/finish/finish-page.component'
+import { EnrolmentPage } from './pages/auth/containers/enrolment-page.component'
+import { AuthService } from './pages/auth/services/auth.service'
+import { ClinicalTasksPage } from './pages/clinical-tasks/containers/clinical-tasks-page.component'
+import { FinishPage } from './pages/finish/containers/finish-page.component'
 import { PrepareDataService } from './pages/finish/services/preparedata.service'
 import { TaskCalendarComponent } from './pages/home/components/task-calendar/task-calendar.component'
 import { TaskInfoComponent } from './pages/home/components/task-info/task-info.component'
 import { TaskProgressComponent } from './pages/home/components/task-progress/task-progress.component'
 import { TickerBarComponent } from './pages/home/components/ticker-bar/ticker-bar.component'
-import { HomePage } from './pages/home/home-page.component'
+import { HomePage } from './pages/home/containers/home-page.component'
 import { InfoScreenComponent } from './pages/questions/components/info-screen/info-screen.component'
 import { AudioInputComponent } from './pages/questions/components/question/audio-input/audio-input.component'
 import { CheckboxInputComponent } from './pages/questions/components/question/checkbox-input/checkbox-input.component'
@@ -45,14 +45,14 @@ import { RadioInputComponent } from './pages/questions/components/question/radio
 import { RangeInputComponent } from './pages/questions/components/question/range-input/range-input.component'
 import { SliderInputComponent } from './pages/questions/components/question/slider-input/slider-input.component'
 import { TimedTestComponent } from './pages/questions/components/question/timed-test/timed-test.component'
-import { QuestionsPage } from './pages/questions/questions-page.component'
+import { QuestionsPage } from './pages/questions/containers/questions-page.component'
 import { AnswerService } from './pages/questions/services/answer.service'
 import { AudioRecordService } from './pages/questions/services/audiorecord.service'
 import { TimeStampService } from './pages/questions/services/timestamp.service'
-import { ReportPage } from './pages/report/report-page.component'
-import { SettingsPage } from './pages/settings/settings-page.component'
-import { SplashPage } from './pages/splash/splash-page.component'
-import { StartPage } from './pages/start/start-page.component'
+import { ReportPage } from './pages/report/containers/report-page.component'
+import { SettingsPage } from './pages/settings/containers/settings-page.component'
+import { SplashPage } from './pages/splash/containers/splash-page.component'
+import { StartPage } from './pages/start/containers/start-page.component'
 import { PipesModule } from './shared/pipes/pipes.module'
 import { TranslatePipe } from './shared/pipes/translate/translate'
 import { AndroidPermissionUtility } from './shared/utilities/android-permission'
