+++ src/app/pages/auth/containers/enrolment-page.component.scss
