+++ src/app/pages/clinical-tasks/containers/clinical-tasks-page.component.scss
