+++ src/app/pages/finish/containers/finish-page.component.scss
