+++ src/app/pages/home/containers/home-page.component.ts
@@ -7,22 +7,21 @@ import {
   Platform
 } from 'ionic-angular'
 
-import { DefaultTask } from '../../../assets/data/defaultConfig'
-import { HomeController } from '../../core/services/home-controller.service'
-import { KafkaService } from '../../core/services/kafka.service'
-import { NotificationService } from '../../core/services/notification.service'
-import { SchedulingService } from '../../core/services/scheduling.service'
-import { StorageService } from '../../core/services/storage.service'
-import { LocKeys } from '../../shared/enums/localisations'
-import { StorageKeys } from '../../shared/enums/storage'
-import { Task, TasksProgress } from '../../shared/models/task'
-import { TranslatePipe } from '../../shared/pipes/translate/translate'
-import { checkTaskIsNow } from '../../shared/utilities/check-task-is-now'
-import { ClinicalTasksPage } from '../clinical-tasks/clinical-tasks-page.component'
-import { QuestionsPage } from '../questions/questions-page.component'
-import { SettingsPage } from '../settings/settings-page.component'
-import { SplashPage } from '../splash/splash-page.component'
-import { StartPage } from '../start/start-page.component'
+import { DefaultTask } from '../../../../assets/data/defaultConfig'
+import { HomeController } from '../../../core/services/home-controller.service'
+import { KafkaService } from '../../../core/services/kafka.service'
+import { NotificationService } from '../../../core/services/notification.service'
+import { SchedulingService } from '../../../core/services/scheduling.service'
+import { StorageService } from '../../../core/services/storage.service'
+import { LocKeys } from '../../../shared/enums/localisations'
+import { StorageKeys } from '../../../shared/enums/storage'
+import { Task, TasksProgress } from '../../../shared/models/task'
+import { TranslatePipe } from '../../../shared/pipes/translate/translate'
+import { checkTaskIsNow } from '../../../shared/utilities/check-task-is-now'
+import { ClinicalTasksPage } from '../../clinical-tasks/containers/clinical-tasks-page.component'
+import { QuestionsPage } from '../../questions/containers/questions-page.component'
+import { SettingsPage } from '../../settings/containers/settings-page.component'
+import { StartPage } from '../../start/containers/start-page.component'
 
 @Component({
   selector: 'page-home',
