+++ src/app/pages/questions/components/question/audio-input/audio-input.component.ts
@@ -14,7 +14,7 @@ import * as opensmile from '../../../../../../../plugins/cordova-plugin-opensmil
 import { Answer } from '../../../../../shared/models/answer'
 import { Section } from '../../../../../shared/models/question'
 import { AndroidPermissionUtility } from '../../../../../shared/utilities/android-permission'
-import { QuestionsPage } from '../../../questions-page.component'
+import { QuestionsPage } from '../../../containers/questions-page.component'
 import { AudioRecordService } from '../../../services/audiorecord.service'
 
 declare var cordova: any
