+++ src/app/pages/splash/containers/splash-page.component.scss
