+++ src/app/pages/start/containers/start-page.component.scss
